+++ src/theme/index.ts
@@ -1,0 +1,2 @@
+export * from './SparkTheme';
+export * from './SparkAppearanceContext';